requirements.txt
fastapi
uvicorn
streamlit 
langchain 
langchain-openai 
langchain-ollama
python-dotenv 
requests 
pandas 
duckdb
sqlglot
langchain_community
langchain_huggingface
sentence_transformers